remove pemberitahuan table

+++ database/migrations/2025_08_25_042702_add_nomor_sk_to_kegiatan_table.php
@@ -12,9 +12,9 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('kegiatans', function (Blueprint $table) {
-            $table->decimal('nomor_sk_tpk')->nullable();
+            $table->unsignedInteger('nomor_sk_tpk')->nullable();
             $table->dateTime('tgl_sk_tpk')->nullable();
-            $table->decimal('nomor_sk_pka')->nullable();
+            $table->unsignedInteger('nomor_sk_pka')->nullable();
             $table->dateTime('tgl_sk_pka')->nullable();
         });
     }
